+++ pkg/webui/containers/fetch-table/index.js
@@ -72,6 +72,7 @@ const m = defineMessages({
     fetchingSearch: base.fetchingSearch,
     pathname: state.router.location.pathname,
     mayAdd: 'mayAdd' in base ? base.mayAdd : true,
+    mayLink: 'mayLink' in base ? base.mayLink : true,
     error: base.error,
   }
 })
@@ -108,6 +109,7 @@ class FetchTable extends Component {
       }),
     ),
     mayAdd: PropTypes.bool,
+    mayLink: PropTypes.bool,
     pageSize: PropTypes.number,
     pathname: PropTypes.string.isRequired,
     rowKeySelector: PropTypes.func,
@@ -134,6 +136,7 @@ class FetchTable extends Component {
     filterValidator,
     itemPathPrefix: '',
     mayAdd: false,
+    mayLink: false,
     searchable: false,
     searchPlaceholderMessage: sharedMessages.search,
     searchQueryMaxLength: 50,
@@ -286,6 +289,7 @@ class FetchTable extends Component {
       fetching,
       fetchingSearch,
       mayAdd,
+      mayLink,
       pageSize,
       addMessage,
       tableTitle,
@@ -382,7 +386,7 @@ class FetchTable extends Component {
             loading={fetching}
             headers={headers}
             rowKeySelector={rowKeySelector}
-            rowHrefSelector={this.rowHrefSelector}
+            rowHrefSelector={mayLink ? this.rowHrefSelector : undefined}
             data={initialFetch ? [] : items}
             emptyMessage={sharedMessages.noMatch}
             handlesPagination={handlesPagination}