+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Added
 
+- Session management page in Account App.
 - Status page references in the Console.
 - Notification Service API that will allow users to receive notifications about their registered entities.
   - This requires a database schema migration (`ttn-lw-stack is-db migrate`) because of the added tables.